refactor routes to be able to refresh react components once network is changed
