+++ crosscash-extension/src/popup/model/constants.ts
@@ -1,6 +1,2 @@
-import { MAINNET } from '../../lib/constants/networks';
-
-// TODO: handle multiple networks from the same provider based on user input
-export const CURRENT_NETWORK = MAINNET;
 export const ETHTransferGasLimit = 21000;
 export const ERC20TransferGasLimit = 65000;
